requirements.txt
pandas
requests
groq
python-dotenv
tabulate
streamlit
